+++ sr25519/Cargo.toml
@@ -14,7 +14,6 @@ defuse-serde-utils.workspace = true
 
 impl-tools.workspace = true
 near-sdk.workspace = true
-serde_with.workspace = true
 
 [dev-dependencies]
 defuse-test-utils.workspace = true
